requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
pycparser==2.19
six==1.12.0
Werkzeug==0.15.4
WTForms==2.2.1
